Refactor for navigability and add nils
